GTM

Documentation for GTM.

GTM.GenerativeTopographicMap

GTM.Posterior

GTM.Responsabilities

GTM.estimateLogLikelihood

GTM.getCoordsMatrix

GTM.getDMatrix

GTM.getGMatrix

GTM.getUMatrix

GTM.getYMatrix

GTM.get_means

GTM.get_modes

GTM.getΦMatrix

GTM.initWMatrix

GTM.initializeVariance

GTM.initβ⁻¹

GTM.updateBeta!

GTM.updateW!

GTM.GenerativeTopographicMap — Method

GTM(k, m, σ, Dataset; α=0.0, tol=0.0001, verbose=false)

Initialize hyperparameters for a GTM model.

k: square root of the number of latent nodes

m: square root of the number of RBF centers in latent space

σ: standard deviation for latent space RBF functions

Dataset: dataset to fit GTM model to. Assumed shape is (n_datapoints, n_features)

α: Weight regularization parameter (0.0 means no regularization)

tol: absolute tolerance used during fitting.

verbose: Set to true for extra print statements.

GTM.Posterior — Method

Posterior(gtm::GenerativeTopographicMap)

Compute a matrix of contributions to posterior probabilities. This is an intermediate result to facilitate computation of true posterior probabilities given by the responsability matrix R. The returned size is (nnodes, ndatapoints). The exp-normalize trick is used for numerical stability.

GTM.Responsabilities — Method

Responsabilities(gtm::GenerativeTopographicMapping, Dataset)

Compute matrix of responsabilities of each node in X to datapoints in Dataset. Return matrix is of size (n_nodes, n_datapoints).

GTM.estimateLogLikelihood — Method

estimateLogLikelihood(gtm::GenerativeTopographicMap, P, Dataset)

Compute the log-likelihood of obtaining our data provided parameters W and β⁻¹.

GTM.getCoordsMatrix — Method

getCoordsMatrix(k::Int)

Generate a matrix of k² node coordinates on a regular grid with $x∈ [-1,1]$ and $y∈[-1, 1]$.

GTM.getDMatrix — Method

getDMatrix(gtm::GenerativeTopographicMap, Dataset)

Compute pairwise distances between projected gaussian centers Y and data points in Dataset. Resulting size is (n_nodes, n_datapoints).

GTM.getGMatrix — Method

getGMatrix(R)

Create diagonal matrix G from responsability matrix R. Return size is (n_nodes, n_nodes).

GTM.getUMatrix — Method

getUMatrix(Dataset)

Perform PCA on the Dataset and return a matrix U containing the first two principal components (first two columns of data covariance matrix) and the variance of the third principal component. Size of returned matrix U is (n_features, 2)

GTM.getYMatrix — Method

getYMatrix(gtm::GenerativeTopographicMap)

Compute Gaussian centers in data space via Y=W*Φ'. Return size is (n_features, n_nodes).

GTM.get_means — Method

get_means(R, X)

Compute responsability weighted mean node position: $⟨x|tⱼ, W, β⟩=Σⱼ Rᵢⱼxᵢ$

GTM.get_modes — Method

get_modes(R, X)

Compute the node corresponding to the mode responsability for each data point.

GTM.getΦMatrix — Method

getΦMatrix(X, M, σ²)

Given a matrix of latent node coordinates X, RBF mean coordinates M, and variance σ², return a matrix Φ of dimension (n_nodes, n_rbf_centers+1). The final column is set to 1.0 to include a bias offset in addition to the RBFs.

GTM.initWMatrix — Method

initWMatrix(X,Φ,U)

Initialize parameter matrix W. Initial weights are chosen so that WΦ' reproduces PCA projections such that WΦ' ≈ UX'

GTM.initializeVariance — Method

initializeVariance(σ::Float64, M::)

Initilize RBF variance by combining supplied standard deviation with minimum RBF mean distance.

GTM.initβ⁻¹ — Method

initβ⁻¹(β⁻¹, Y)

Initialized β⁻¹ using our first guess for β⁻¹ (from 3rd principal component variance) and the mean distance between projected rbf centers in data space.

GTM.updateBeta! — Method

updateBeta(R, D)

GTM.updateW! — Method

updateW!(gtm::GenerativeTopographicMap, Dataset)

Update model weights using responsability matrix.